Dev : failing build and refactored more
