+++ packages/arb-token-bridge-ui/src/util/TokenUtils.ts
@@ -1,20 +1,9 @@
-import Ajv from 'ajv'
-import addFormats from 'ajv-formats'
-import { schema, TokenList } from '@uniswap/token-lists'
 import { BigNumber, constants } from 'ethers'
 import { Provider } from '@ethersproject/providers'
 import { Erc20Bridger, MultiCaller } from '@arbitrum/sdk'
 import { StandardArbERC20__factory } from '@arbitrum/sdk/dist/lib/abi/factories/StandardArbERC20__factory'
 import { ERC20__factory } from '@arbitrum/sdk/dist/lib/abi/factories/ERC20__factory'
 import { L1TokenData, L2TokenData } from '../token-bridge-sdk/index'
-
-export const validateTokenList = (tokenList: TokenList) => {
-  const ajv = new Ajv()
-  addFormats(ajv)
-  const validate = ajv.compile(schema)
-
-  return validate(tokenList)
-}
 
 export function getDefaultTokenName(address: string) {
   const lowercased = address.toLowerCase()
